test(query-runner): split rename-constraint tests per primitive and fix memory-down flow
